# RISC-V 64: Add assembly code for SHA-256