# Description

Move common defines out of AES file to header file.

# Testing

Regression tested RISC-V 64 ASM.

# Checklist

 - [ ] added tests
 - [ ] updated/added doxygen
 - [ ] updated appropriate READMEs
 - [ ] Updated manual and documentation