@dgarske 2024-07-22
@SparkiDev I will run this on actual hardware and compare performance for SHA2-256.