@dgarske 2024-07-22
HiFive Unleashed at 1.4GHz:

SHA-256: 18.326 MiB/s -> 20.399 MiB/s

PR 7758 (with RISCV-ASM):

```
root@HiFiveU:~/wolfssl-riscv# ./configure --enable-riscv-asm && make
root@HiFiveU:~/wolfssl-riscv# ./wolfcrypt/benchmark/benchmark
------------------------------------------------------------------------------
 wolfSSL version 5.7.2
------------------------------------------------------------------------------
Math: 	Multi-Precision: Wolf(SP) word-size=64 bits=3072 sp_int.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
RNG                         10 MiB took 1.094 seconds,    9.141 MiB/s
AES-128-CBC-enc             20 MiB took 1.062 seconds,   18.837 MiB/s
AES-128-CBC-dec             20 MiB took 1.059 seconds,   18.893 MiB/s
AES-192-CBC-enc             20 MiB took 1.240 seconds,   16.130 MiB/s
AES-192-CBC-dec             20 MiB took 1.233 seconds,   16.216 MiB/s
AES-256-CBC-enc             15 MiB took 1.046 seconds,   14.343 MiB/s
AES-256-CBC-dec             15 MiB took 1.039 seconds,   14.440 MiB/s
AES-128-GCM-enc             15 MiB took 1.301 seconds,   11.528 MiB/s
AES-128-GCM-dec             15 MiB took 1.300 seconds,   11.542 MiB/s
AES-192-GCM-enc             15 MiB took 1.423 seconds,   10.543 MiB/s
AES-192-GCM-dec             15 MiB took 1.425 seconds,   10.528 MiB/s
AES-256-GCM-enc             10 MiB took 1.025 seconds,    9.755 MiB/s
AES-256-GCM-dec             10 MiB took 1.025 seconds,    9.760 MiB/s
GMAC Table 4-bit            31 MiB took 1.025 seconds,   30.245 MiB/s
CHACHA                      35 MiB took 1.128 seconds,   31.018 MiB/s
CHA-POLY                    25 MiB took 1.095 seconds,   22.825 MiB/s
MD5                         75 MiB took 1.011 seconds,   74.202 MiB/s
POLY1305                    90 MiB took 1.044 seconds,   86.186 MiB/s
SHA                         35 MiB took 1.064 seconds,   32.891 MiB/s
SHA-256                     25 MiB took 1.226 seconds,   20.399 MiB/s
SHA-384                     25 MiB took 1.128 seconds,   22.153 MiB/s
SHA-512                     25 MiB took 1.129 seconds,   22.146 MiB/s
SHA-512/224                 25 MiB took 1.128 seconds,   22.163 MiB/s
SHA-512/256                 25 MiB took 1.129 seconds,   22.149 MiB/s
HMAC-MD5                    75 MiB took 1.010 seconds,   74.237 MiB/s
HMAC-SHA                    35 MiB took 1.066 seconds,   32.830 MiB/s
HMAC-SHA256                 25 MiB took 1.223 seconds,   20.446 MiB/s
HMAC-SHA384                 25 MiB took 1.128 seconds,   22.165 MiB/s
HMAC-SHA512                 25 MiB took 1.128 seconds,   22.155 MiB/s
PBKDF2                       3 KiB took 1.007 seconds,    2.544 KiB/s
RSA     2048   public      1500 ops took 1.047 sec, avg 0.698 ms, 1432.120 ops/sec
RSA     2048  private       100 ops took 4.375 sec, avg 43.749 ms, 22.858 ops/sec
DH      2048  key gen       116 ops took 1.006 sec, avg 8.671 ms, 115.326 ops/sec
DH      2048    agree       100 ops took 1.843 sec, avg 18.430 ms, 54.259 ops/sec
ECC   [      SECP256R1]   256  key gen       200 ops took 1.282 sec, avg 6.409 ms, 156.031 ops/sec
ECDHE [      SECP256R1]   256    agree       200 ops took 1.277 sec, avg 6.386 ms, 156.591 ops/sec
ECDSA [      SECP256R1]   256     sign       200 ops took 1.310 sec, avg 6.549 ms, 152.697 ops/sec
ECDSA [      SECP256R1]   256   verify       300 ops took 1.330 sec, avg 4.432 ms, 225.626 ops/sec
Benchmark complete
```

Master (with RISCV-ASM):
```
root@HiFiveU:~/wolfssl-riscv# ./configure --enable-riscv-asm && make
root@HiFiveU:~/wolfssl-riscv# ./wolfcrypt/benchmark/benchmark
------------------------------------------------------------------------------
 wolfSSL version 5.7.2
------------------------------------------------------------------------------
Math: 	Multi-Precision: Wolf(SP) word-size=64 bits=3072 sp_int.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
RNG                         10 MiB took 1.366 seconds,    7.320 MiB/s
AES-128-CBC-enc             20 MiB took 1.070 seconds,   18.688 MiB/s
AES-128-CBC-dec             20 MiB took 1.065 seconds,   18.775 MiB/s
AES-192-CBC-enc             20 MiB took 1.240 seconds,   16.134 MiB/s
AES-192-CBC-dec             20 MiB took 1.231 seconds,   16.244 MiB/s
AES-256-CBC-enc             15 MiB took 1.052 seconds,   14.256 MiB/s
AES-256-CBC-dec             15 MiB took 1.045 seconds,   14.359 MiB/s
AES-128-GCM-enc             15 MiB took 1.303 seconds,   11.511 MiB/s
AES-128-GCM-dec             15 MiB took 1.303 seconds,   11.508 MiB/s
AES-192-GCM-enc             15 MiB took 1.431 seconds,   10.479 MiB/s
AES-192-GCM-dec             15 MiB took 1.432 seconds,   10.475 MiB/s
AES-256-GCM-enc             10 MiB took 1.034 seconds,    9.669 MiB/s
AES-256-GCM-dec             10 MiB took 1.034 seconds,    9.669 MiB/s
GMAC Table 4-bit            31 MiB took 1.003 seconds,   30.917 MiB/s
CHACHA                      35 MiB took 1.139 seconds,   30.740 MiB/s
CHA-POLY                    25 MiB took 1.103 seconds,   22.666 MiB/s
MD5                         75 MiB took 1.009 seconds,   74.360 MiB/s
POLY1305                    90 MiB took 1.041 seconds,   86.473 MiB/s
SHA                         35 MiB took 1.061 seconds,   32.977 MiB/s
SHA-256                     20 MiB took 1.091 seconds,   18.326 MiB/s
SHA-384                     25 MiB took 1.152 seconds,   21.699 MiB/s
SHA-512                     25 MiB took 1.153 seconds,   21.689 MiB/s
SHA-512/224                 25 MiB took 1.151 seconds,   21.721 MiB/s
SHA-512/256                 25 MiB took 1.151 seconds,   21.718 MiB/s
HMAC-MD5                    75 MiB took 1.009 seconds,   74.366 MiB/s
HMAC-SHA                    35 MiB took 1.061 seconds,   32.977 MiB/s
HMAC-SHA256                 20 MiB took 1.093 seconds,   18.302 MiB/s
HMAC-SHA384                 25 MiB took 1.151 seconds,   21.725 MiB/s
HMAC-SHA512                 25 MiB took 1.151 seconds,   21.723 MiB/s
PBKDF2                       2 KiB took 1.001 seconds,    2.279 KiB/s
RSA     2048   public      1500 ops took 1.059 sec, avg 0.706 ms, 1416.165 ops/sec
RSA     2048  private       100 ops took 4.409 sec, avg 44.086 ms, 22.683 ops/sec
DH      2048  key gen       115 ops took 1.003 sec, avg 8.718 ms, 114.703 ops/sec
DH      2048    agree       100 ops took 1.845 sec, avg 18.454 ms, 54.188 ops/sec
ECC   [      SECP256R1]   256  key gen       200 ops took 1.278 sec, avg 6.390 ms, 156.485 ops/sec
ECDHE [      SECP256R1]   256    agree       200 ops took 1.273 sec, avg 6.365 ms, 157.107 ops/sec
ECDSA [      SECP256R1]   256     sign       200 ops took 1.308 sec, avg 6.542 ms, 152.860 ops/sec
ECDSA [      SECP256R1]   256   verify       300 ops took 1.337 sec, avg 4.456 ms, 224.392 ops/sec
```