@SparkiDev 2024-07-25
@dgarske,

Please generate benchmarks with assembly code again.

Thanks,
Sean